+++ src/main/java/com/org/bank/dao/ClassDTOMapper.java
@@ -20,7 +20,7 @@ public interface ClassDTOMapper {
 
     List<ClassDTO> selectList(@Param("record") ClassDTO record, @Param("pager")Pager pager);
 
-    Integer countTotal(ClassDTO record);
+    Long countTotal(ClassDTO record);
 
     int updateByPrimaryKeySelective(ClassDTO record);
